deserialization implementation and other changes
Implemented deserialize method: a string object is read from the file in which it was serialized and printed
Modified serialize method: serialization makes sense to store an object to a file. The file where the serialized object has to be saved is provided along with the string to serialize. The string is then serialized and stored in such file.